Create and populate forecast fragment

+++ app/src/main/java/com/lukakordic/weatherapp/data/model/Wind.kt
@@ -1,8 +1,5 @@
 package com.lukakordic.weatherapp.data.model
 
-import android.arch.persistence.room.Entity
-
-//@Entity
 data class Wind(val speed: Double,
-                val deg: Int
+                val deg: Double
 )
